integrate new parser into sim + unsafety checking
I'm too lazy to split this commit...

The new parser should now work with the simulation backend. Unsafety
checking is added as a frame by frame checking, and doesn't influence
the simulation. Currently only `assert`s in the controller itself is
checked, but other options can be added with ease. When plotted with
plotly, the unsafety set is denoted by segments with a black color

Tested on `demo3` and `ball_bounces_dev` examples